requirements.txt
google-cloud-translate
argparse